First algorithm to generate lattice (redoing now) More compilation improvements

+++ src/interface/IsingModel.h
@@ -70,7 +70,7 @@ class IsingModel {
         typedef struct {
            int S;
            bool active;
-           std::vector<int> coords;
+           std::vector<double> coords;
         } spin;
 
         // Settings
